chore: tests: remove OpenSUSE Leap 15.0
We're hitting this bug: https://bugzilla.opensuse.org/show_bug.cgi?id=1146027
Hence the tests randomly fail because of that. The workaround is not really clean
(installing a custom-built package), so we're pulling OpenSUSE Leap 15.0 from our
automated tests workflow. Note that it's still supported, the tests can be launched
manually, and bugs under OpenSUSE Leap 15.0 will still be fixed.