chore: faster tests by removing grant/revoke command dance
When restricted commands need to be used during tests,
we now use "account0" which has all these commands granted,
instead of granting/revoking commands every time with no added
value with respect to the tests.

This was previously required for OSes that have a limit to the
number of groups an account can be a member of, but these OSes
have now long been unsupported.