enh: clush: keep hosts order and no longer dedupe them
The hosts order might have some importance for the caller, so don't
attempt to sort them. Also don't try to dedupe them, it might be
for some reason that the same host appears twice in the list,
so don't make any assumptions on how we are used and just apply
the command on the given hosts, exactly in the order given.